forRemoval警告の既存違反を解消する
